gcs: Move TTG CUPS to GCS
