+++ pkg/gatewayconfigurationserver/gcsv2/server.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cups
+package gcsv2
 
 import (
 	"context"
